+++ R/utils.R
@@ -1,7 +1,3 @@
-on_windows <- function() {
-  pkgenv[["on_windows"]]
-}
-
 `%||%` <- function(x, y) {
   if (is.null(x)) {
     y