+++ R/zzz.R
@@ -1,5 +1,0 @@
-pkgenv <- new.env(parent = emptyenv())
-
-.onLoad <- function(libname, pkgname) {
-  populate_pkgenv() # nocov
-}