+++ tests/testthat/test-utils.R
@@ -1,7 +1,3 @@
-test_that("on_windows()", {
-  expect_true(is.logical(on_windows()))
-})
-
 test_that("%||%", {
   expect_equal("x" %||% "y", "x")
   expect_equal(character(0) %||% "y", character(0))