+++ NEWS.md
@@ -1,6 +1,6 @@
 # proffer 0.0.1.9000
 
-
+* Use `processx` instead of `callr` (#5, @gaborcsardi).
 
 # proffer 0.0.1